Adding Pokemon File and PokemonTester File
Created a file that creates a pokemon class with different attributes to the pokemon.
The pokemon class has a constructor that takes in the name, type, and level of the pokemon
or you can just call the class as an object and input values indiviidually.

PokemonTester test the pokemon class and prints out the values of the pokemons created through various methods of constructing the object.